pyproject.toml
[project]
name = "coloraria"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "google-genai>=1.55.0",
  "neo4j==6.0.3",
  "memory-profiler==0.61.0",
  "psutil==7.1.3",
  "pydantic>=2.12.3",
  "pydantic-core==2.41.4",
  "pyjwt>=2.8.0",
  "python-dotenv==1.2.1",
  "pyyaml==6.0.3",
  "dotenv==0.9.9",
  "fastapi==0.124.2",
  "httpcore==1.0.9",
  "httpx==0.28.1",
  "h11==0.16.0",
  "idna==3.11",
  "requests==2.32.5",
  "uvicorn==0.38.0",
  "lxml>=6.0.2",
]

[dependency-groups]
ingestion = [
]

inference = [
    "langchain-core >= 1.1.3",
    "langchain-google-genai >= 4.0.0",
    "langchain-openai >= 1.1.3",
    "langgraph >= 1.0.4",
    "langgraph-checkpoint >= 3.0.1",
    "langgraph-checkpoint-sqlite >= 3.0.1",
    "langgraph-checkpoint-mysql[pymysql] >= 2.0.0",
    "langgraph-prebuilt >= 1.0.5",
    "langgraph-sdk >= 0.2.15",
    "langsmith >= 0.4.59",
    "arize-phoenix>=5.0.0",
    "openinference-instrumentation-langchain>=0.1.0",
    "openinference-instrumentation-google-genai>=0.1.0",
    "openinference-instrumentation-openai>=0.1.0",
    "opentelemetry-sdk>=1.20.0",
    "opentelemetry-api>=1.20.0",
    "sqlalchemy>=2.0.0",
    "pymysql>=1.1.0",
    "bcrypt>=4.0.0",
    "python-bcrypt==0.3.2",


]
